refactor code actions file
