+++ src/StaticLS/IDE/CodeActions/Types.hs
@@ -7,6 +7,7 @@ import Data.Path (AbsPath)
 import Data.Text
 import Language.LSP.Protocol.Types (Range (..), TextDocumentIdentifier (..))
 import StaticLS.StaticLsEnv
+import StaticLS.IDE.SourceEdit (SourceEdit)
 
 data Context = Context
   { textDocument :: !TextDocumentIdentifier
@@ -26,6 +27,25 @@ data GlobalCodeAction = GlobalCodeAction
   { name :: !Text
   , run :: Context -> StaticLsM (Maybe ())
   }
-  
+
 $(deriveJSON defaultOptions ''CodeActionMessageKind)
 $(deriveJSON defaultOptions ''CodeActionMessage)
+
+data Assist = Assist
+  { label :: !Text
+  , sourceEdit :: Either SourceEdit CodeActionMessage
+  }
+
+mkAssist :: Text -> SourceEdit -> Assist
+mkAssist label sourceEdit =
+  Assist
+    { label
+    , sourceEdit = Left sourceEdit
+    }
+
+mkLazyAssist :: Text -> CodeActionMessage -> Assist
+mkLazyAssist label sourceEdit =
+  Assist
+    { label
+    , sourceEdit = Right sourceEdit
+    }
